+++ CHANGES.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `load_best` attribute to `EarlyStopping` callback to automatically load module weights of the best result at the end of training
 - Added a method, `trim_for_prediction`, on the net classes, which trims the net from everything not required for using it for prediction; call this after fitting to reduce the size of the net
 - Added experimental support for [huggingface accelerate](https://github.com/huggingface/accelerate); use the provided mixin class to add advanced training capabilities provided by the accelerate library to skorch
+- Added possibility for `dirname` argument of `Checkpoint` to be a callable that returns the actual directory name; this can be used for generating directory names dynamically, e.g. in case multiple checkpoints are being used in parallel training (#848)
 
 ### Changed
 - The minimum required scikit-learn version has been bumped to 0.22.0